Add: [NewGRF] Second vehicle property for additional callback flags.

+++ src/engine_type.h
@@ -150,10 +150,10 @@ struct EngineInfo {
 	CargoID cargo_type;
 	CargoTypes refit_mask;
 	byte refit_cost;
-	byte misc_flags;    ///< Miscellaneous flags. @see EngineMiscFlags
-	byte callback_mask; ///< Bitmask of vehicle callbacks that have to be called
-	int8 retire_early;  ///< Number of years early to retire vehicle
-	StringID string_id; ///< Default name of engine
+	byte misc_flags;         ///< Miscellaneous flags. @see EngineMiscFlags
+	uint16 callback_mask;    ///< Bitmask of vehicle callbacks that have to be called
+	int8 retire_early;       ///< Number of years early to retire vehicle
+	StringID string_id;      ///< Default name of engine
 	uint16 cargo_age_period; ///< Number of ticks before carried cargo is aged.
 	EngineID variant_id;     ///< Engine variant ID. If set, will be treated specially in purchase lists.
 	ExtraEngineFlags extra_flags;
